Only show admintools.conf debug in dev mode
This hides some chatty messages from the operator log. We normally would
dump out the contents of the admintools.conf. But this is suppressed,
unless the devMode flag is set in the operator.

Changes to the operator main.go to flow down the entire config struct
(now called OperatorConfig).
